timer for game added

+++ hat_game/app/src/main/java/com/example/hat/entity/GameSettings.kt
@@ -4,7 +4,7 @@ import java.io.Serializable
 
 data class GameSettings(
     var vocabular: MutableList<String>? = null,
-    var secondsPerStep: Int? = null,
+    var secondsPerStep: Long? = null,
     var team1: String? = null,
     var team2: String? = null
 ) : Serializable
